+++ src/docstring2tsx/__main__.py
@@ -10,6 +10,7 @@ src_dir = Path(__file__).parent.parent.parent / "src"
 sys.path.insert(0, str(src_dir))
 
 from docstring_2tsx.converter import file_to_tsx
+
 from utils.shared import process_package
 
 # Configure logging
