+++ src/docstring2tsx/cli.py
@@ -4,6 +4,7 @@ import argparse
 from pathlib import Path
 
 from docstring_2tsx.converter import file_to_tsx
+
 from utils.importer import import_module_from_file
 from utils.shared import process_package
