+++ src/docstring2tsx/utils/shared.py
