+++ src/docstring2tsx/utils/signature_formatter.py
